[finish #163934346] created the base directory structure for v2
